+++ src/main/java/Entitys/BankInterface.java
@@ -6,9 +6,9 @@ import java.util.List;
  * Created by carlosmorais on 15/02/16.
  */
 public interface BankInterface {
-    int getBalance(long accountId);
-    long newAccount(String opId);
-    boolean mov(long accountId, int amount, String opId);
-    boolean transf(long source, long dest, int amount, String opId);
-    List<Movement> movList(long id, int n);
+    int getBalance(int accountId);
+    int newAccount(String opId);
+    boolean mov(int accountId, int amount, String opId);
+    boolean transf(int source, int dest, int amount, String opId);
+    List<Movement> movList(int id, int n);
 }
